# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API disbursement dengan idempotency, approval yang aman terhadap race condition, soft delete, audit trail terpisah, dan structured logging dengan propagasi request ID.",
        "title": "Disbursement API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/disbursements": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "disbursements"
                ],
                "summary": "List disbursements",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number, default 1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size, default 20, max 100",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Matches recipient_name or account_number",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "PENDING, APPROVED, or REJECTED",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "RFC3339 lower bound on created_at",
                        "name": "date_from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "RFC3339 upper bound on created_at",
                        "name": "date_to",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Whitelisted column, e.g. created_at, amount",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "asc or desc",
                        "name": "sort_order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/disbursement.Response"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/definitions/response.PageMeta"
                                        }
                                    }
                                }
                            ]
                        },
                        "headers": {
                            "X-Request-ID": {
                                "type": "string",
                                "description": "Request correlation ID; echoed from the request header or generated if absent"
                            }
                        }
                    },
                    "400": {
                        "description": "VALIDATION_ERROR: sort_by/sort_order not in whitelist, or invalid date",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "401": {
                        "description": "UNAUTHORIZED",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "disbursements"
                ],
                "summary": "Create disbursement",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID; a retried request with the same key and body within 24h replays the first response",
                        "name": "Idempotency-Key",
                        "in": "header"
                    },
                    {
                        "description": "Disbursement details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/disbursement.CreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/disbursement.Response"
                                        }
                                    }
                                }
                            ]
                        },
                        "headers": {
                            "X-Idempotent-Replayed": {
                                "type": "string",
                                "description": "true when this response is a replay of an earlier request with the same Idempotency-Key"
                            },
                            "X-Request-ID": {
                                "type": "string",
                                "description": "Request correlation ID; echoed from the request header or generated if absent"
                            }
                        }
                    },
                    "400": {
                        "description": "VALIDATION_ERROR: missing field or amount below minimum",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "401": {
                        "description": "UNAUTHORIZED: missing or invalid access token",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "409": {
                        "description": "IDEMPOTENCY_IN_PROGRESS: a request with this key is still processing",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "422": {
                        "description": "IDEMPOTENCY_KEY_REUSED: same key, different request body",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "429": {
                        "description": "RATE_LIMITED",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    }
                }
            }
        },
        "/disbursements/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "disbursements"
                ],
                "summary": "Get disbursement",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Disbursement ID, e.g. DSB-000001",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/disbursement.Response"
                                        }
                                    }
                                }
                            ]
                        },
                        "headers": {
                            "X-Request-ID": {
                                "type": "string",
                                "description": "Request correlation ID; echoed from the request header or generated if absent"
                            }
                        }
                    },
                    "401": {
                        "description": "UNAUTHORIZED",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "404": {
                        "description": "NOT_FOUND: unknown ID or soft-deleted",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "disbursements"
                ],
                "summary": "Soft-delete a disbursement",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Disbursement ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content",
                        "headers": {
                            "X-Request-ID": {
                                "type": "string",
                                "description": "Request correlation ID; echoed from the request header or generated if absent"
                            }
                        }
                    },
                    "401": {
                        "description": "UNAUTHORIZED",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "403": {
                        "description": "FORBIDDEN: caller is not superadmin",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "404": {
                        "description": "NOT_FOUND: unknown ID or already soft-deleted",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "409": {
                        "description": "CONFLICT: disbursement is no longer PENDING",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    }
                }
            }
        },
        "/disbursements/{id}/status": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "disbursements"
                ],
                "summary": "Approve or reject a disbursement",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Disbursement ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New status: APPROVED or REJECTED",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/disbursement.UpdateStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/disbursement.Response"
                                        }
                                    }
                                }
                            ]
                        },
                        "headers": {
                            "X-Request-ID": {
                                "type": "string",
                                "description": "Request correlation ID; echoed from the request header or generated if absent"
                            }
                        }
                    },
                    "400": {
                        "description": "VALIDATION_ERROR: status is neither APPROVED nor REJECTED",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "401": {
                        "description": "UNAUTHORIZED",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "403": {
                        "description": "FORBIDDEN: caller is not admin or superadmin",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "404": {
                        "description": "NOT_FOUND: unknown ID or soft-deleted",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "409": {
                        "description": "CONFLICT: disbursement is no longer PENDING",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "disbursement.CreateRequest": {
            "type": "object",
            "properties": {
                "account_number": {
                    "type": "string"
                },
                "amount": {
                    "type": "integer"
                },
                "bank_code": {
                    "type": "string"
                },
                "note": {
                    "type": "string"
                },
                "recipient_name": {
                    "type": "string"
                }
            }
        },
        "disbursement.Response": {
            "type": "object",
            "properties": {
                "account_number": {
                    "type": "string"
                },
                "admin_fee": {
                    "type": "integer"
                },
                "amount": {
                    "type": "integer"
                },
                "approved_by": {
                    "type": "string"
                },
                "bank_code": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "note": {
                    "type": "string"
                },
                "recipient_name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "disbursement.UpdateStatusRequest": {
            "type": "object",
            "properties": {
                "note": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "response.Envelope": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/response.ErrorBody"
                },
                "meta": {},
                "success": {
                    "type": "boolean"
                }
            }
        },
        "response.ErrorBody": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "response.PageMeta": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Ketik \"Bearer\" diikuti spasi dan access token, contoh: \"Bearer eyJhbGciOi...\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}